+++ Sources/Domain/SpotPost/Components/SpotPostSubmitButton.swift
@@ -1,24 +1,62 @@
 import Foundation
 import SwiftUI
+import CoreLocation
 
 public struct SpotPostSubmitButton: View {
-    let isDisabled: Bool
+    @Environment(\.presentationMode) var presentationMode
+
+    @StateObject var upload = Upload()
+    @StateObject var mutation = Mutation<SpotAddMutation>()
+
+    @State var error: IdentifiableError?
+
+    @Binding var image: UIImage?
+    @Binding var title: String
+    @Binding var geoPoint: CLLocationCoordinate2D?
+
+    var submitButtonIsDisabled: Bool {
+        image == nil || title.isEmpty || geoPoint == nil
+    }
 
     public var body: some View {
         Button(
-            action: {
-                // TODO:
-            },
+            action: save,
             label: {
                 Text("保存")
                     .foregroundColor(.white)
                     .font(.body)
                     .fontWeight(.medium)
             })
-            .disabled(isDisabled)
-            .frame(width: 200, height: 44, alignment: .center)
-            .background(isDisabled ? Color.disabled.gradient : GradientColor.barn
-            )
+            .disabled(submitButtonIsDisabled)
+            .buttonStyle(PrimaryButtonStyle(isLoading: mutation.isProcessing))
+            .alert(item: $error) { error in
+                Alert(title: Text("エラーが発生しました"), message: Text(error.localizedDescription), dismissButton: .default(Text("OK")))
+            }
+    }
+
+    private func save() {
+        guard let image = image, let geoPoint = geoPoint else {
+            return
+        }
+        Task {
+            // TODO:
+            do {
+                let uploaded = try await upload(path: .spot(userID: "", spotID: ""), image: image)
+                try await mutation(
+                    for: .init(
+                        spotAddInput: .init(
+                            title: title,
+                            imageUrl: uploaded.url,
+                            latitude: geoPoint.latitude,
+                            longitude: geoPoint.longitude
+                        )
+                    )
+                )
+                presentationMode.wrappedValue.dismiss()
+            } catch {
+                self.error = .init(error)
+            }
+        }
     }
 }
