+++ Sources/Domain/SpotPost/SpotPostView.swift
@@ -9,17 +9,10 @@ import FirebaseStorageSwift
 struct SpotPostView: View {
     @Environment(\.presentationMode) var presentationMode
 
-    @StateObject var upload = Upload()
-    @StateObject var mutation = Mutation<SpotAddMutation>()
-
     @State var error: Error?
     @State var image: UIImage?
     @State var title: String = ""
     @State var geoPoint: CLLocationCoordinate2D?
-
-    var submitButtonIsDisabled: Bool {
-        image == nil || title.isEmpty || geoPoint == nil
-    }
 
     var body: some View {
         NavigationView {
@@ -31,17 +24,11 @@ struct SpotPostView: View {
                         SpotPostTitle(title: $title)
                         SpotPostGeoPoint(geoPoint: $geoPoint)
                         Spacer()
-                        PrimaryButton(
-                            isLoading: mutation.isProcessing,
-                            action: save,
-                            label: {
-                                Text("保存")
-                                    .foregroundColor(.white)
-                                    .font(.body)
-                                    .fontWeight(.medium)
-                            }
+                        SpotPostSubmitButton(
+                            image: $image,
+                            title: $title,
+                            geoPoint: $geoPoint
                         )
-                        SpotPostSubmitButton(isDisabled: submitButtonIsDisabled)
                         Spacer().frame(height: 32)
                     }
                 }
@@ -61,30 +48,6 @@ struct SpotPostView: View {
         }
     }
 
-    private func save() {
-        guard let image = image, let geoPoint = geoPoint else {
-            return
-        }
-        Task {
-            // TODO:
-            do {
-                let uploaded = try await upload(path: .spot(userID: "", spotID: ""), image: image)
-                try await mutation(
-                    for: .init(
-                        spotAddInput: .init(
-                            title: title,
-                            imageUrl: uploaded.url,
-                            latitude: geoPoint.latitude,
-                            longitude: geoPoint.longitude
-                        )
-                    )
-                )
-                presentationMode.wrappedValue.dismiss()
-            } catch {
-                self.error = error
-            }
-        }
-    }
 
 }
